// book-network-ui/src/app/services/models/page-response-book-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BookResponse } from '../models/book-response';

export interface PageResponseBookResponse {
  content?: Array<BookResponse>;
  first?: boolean;
  last?: boolean;
  page?: number;
  size?: number;
  total?: number;
  totalPages?: number;
}